backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
langchain==0.1.0
langchain-community==0.0.10
pinecone-client==3.0.0
boto3==1.34.0
sagemaker==2.199.0
PyPDF2==3.0.1
python-multipart==0.0.6
pydantic
numpy==1.24.3
sentence-transformers==2.2.2
python-dotenv==1.0.0